Move index sub objects to index_subobjects
