+++ yt/frontends/boxlib/data_structures.py
@@ -7,7 +7,7 @@ from stat import ST_CTIME
 
 import numpy as np
 
-from yt.data_objects.grid_patch import AMRGridPatch
+from yt.data_objects.index_subobjects.grid_patch import AMRGridPatch
 from yt.data_objects.static_output import Dataset
 from yt.funcs import ensure_tuple, mylog, setdefaultattr
 from yt.geometry.grid_geometry_handler import GridIndex
