+++ yt/frontends/gdf/data_structures.py
@@ -3,7 +3,7 @@ import weakref
 
 import numpy as np
 
-from yt.data_objects.grid_patch import AMRGridPatch
+from yt.data_objects.index_subobjects.grid_patch import AMRGridPatch
 from yt.data_objects.static_output import Dataset
 from yt.funcs import ensure_tuple, just_one, setdefaultattr
 from yt.geometry.grid_geometry_handler import GridIndex
