+++ yt/frontends/stream/data_structures.py
@@ -9,12 +9,15 @@ from numbers import Number as numeric_type
 import numpy as np
 
 from yt.data_objects.field_data import YTFieldData
-from yt.data_objects.grid_patch import AMRGridPatch
-from yt.data_objects.octree_subset import OctreeSubset
+from yt.data_objects.index_subobjects.grid_patch import AMRGridPatch
+from yt.data_objects.index_subobjects.octree_subset import OctreeSubset
+from yt.data_objects.index_subobjects.unstructured_mesh import (
+    SemiStructuredMesh,
+    UnstructuredMesh,
+)
 from yt.data_objects.particle_unions import ParticleUnion
 from yt.data_objects.static_output import Dataset, ParticleFile
 from yt.data_objects.unions import MeshUnion
-from yt.data_objects.unstructured_mesh import SemiStructuredMesh, UnstructuredMesh
 from yt.frontends.exodus_ii.util import get_num_pseudo_dims
 from yt.frontends.sph.data_structures import SPHParticleIndex
 from yt.funcs import ensure_list, issue_deprecation_warning, iterable
