+++ yt/frontends/ytdata/data_structures.py
@@ -6,7 +6,7 @@ from numbers import Number as numeric_type
 import numpy as np
 
 from yt.data_objects.data_containers import GenerationInProgress
-from yt.data_objects.grid_patch import AMRGridPatch
+from yt.data_objects.index_subobjects.grid_patch import AMRGridPatch
 from yt.data_objects.particle_unions import ParticleUnion
 from yt.data_objects.profiles import (
     Profile1DFromDataset,
